Move vespa search continuation token tests to separate file
